% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggy_gap.R
\name{gg.y_gap}
\alias{gg.y_gap}
\title{gg.y_gap}
\usage{
gg.y_gap(
  p,
  y_segment_start,
  y_segment_end,
  break_step,
  gap_lines = "rl",
  gap_line_length = 0.02,
  gap_width = 0.01,
  x_min = NULL
)
}
\arguments{
\item{p}{your ggplot}

\item{y_segment_start}{minimum y-value (numerical; e.g., 50)}

\item{y_segment_end}{maximum y-value (numerical; e.g., 100)}

\item{break_step}{step size between axis-ticks (numerical; e.g., 10)}

\item{gap_lines}{gap lines to the right/ left/ both (character string; i.e., "r", "l", "rl")}

\item{gap_line_length}{length of gap lines in percent (float; e.g., 0.02)}

\item{gap_width}{width of the gap in percent (float; e.g., 0.01)}

\item{x_min}{x-axis start value (numerical; e.g., 0)}
}
\value{
your ggplot
}
\description{
Create a gap in your ggplot y-axis.
}
\examples{
gg.y_gap(p, 50, 100, 10)
gg.y_gap(p, 50, 100, 10, "r", 0.05, 0.04, 0)
}
